refactor(data): move tempMovieData and tempWatchedData into separate files
