+++ src/components/App.tsx
@@ -1,51 +1,6 @@
 import { useState } from "react";
-
-const tempMovieData = [
-  {
-    imdbID: "tt0468569",
-    title: "The Dark Knight",
-    year: "2008",
-    poster:
-      "https://m.media-amazon.com/images/M/MV5BMTMxNTMwODM0NF5BMl5BanBnXkFtZTcwODAyMTk2Mw@@._V1_.jpg",
-  },
-  {
-    imdbID: "tt0107048",
-    title: "Groundhog Day",
-    year: "1993",
-    poster:
-      "https://m.media-amazon.com/images/M/MV5BZWIxNzM5YzQtY2FmMS00Yjc3LWI1ZjUtNGVjMjMzZTIxZTIxXkEyXkFqcGdeQXVyNjU0OTQ0OTY@._V1_.jpg",
-  },
-  {
-    imdbID: "tt1745960",
-    title: "Top Gun: Maverick",
-    year: "2022",
-    poster:
-      "https://m.media-amazon.com/images/M/MV5BZWYzOGEwNTgtNWU3NS00ZTQ0LWJkODUtMmVhMjIwMjA1ZmQwXkEyXkFqcGdeQXVyMjkwOTAyMDU@._V1_.jpg",
-  },
-];
-
-const tempWatchedData = [
-  {
-    imdbID: "tt0468569",
-    title: "The Dark Knight",
-    year: "2008",
-    poster:
-      "https://m.media-amazon.com/images/M/MV5BMTMxNTMwODM0NF5BMl5BanBnXkFtZTcwODAyMTk2Mw@@._V1_.jpg",
-    runtime: 152,
-    imdbRating: 9.0,
-    userRating: 9.9,
-  },
-  {
-    imdbID: "tt1745960",
-    title: "Top Gun: Maverick",
-    year: "2022",
-    poster:
-      "https://m.media-amazon.com/images/M/MV5BZWYzOGEwNTgtNWU3NS00ZTQ0LWJkODUtMmVhMjIwMjA1ZmQwXkEyXkFqcGdeQXVyMjkwOTAyMDU@._V1_.jpg",
-    runtime: 130,
-    imdbRating: 8.2,
-    userRating: 9,
-  },
-];
+import tempMovieData from "../data/tempMovieData";
+import tempWatchedData from "../data/tempWatchedData";
 
 const average = (arr: number[]): number => {
   const sum = arr.reduce((acc, cur) => acc + cur, 0);
