+++ src/data/tempWatchedData.tsx
@@ -1,0 +1,24 @@
+const tempWatchedData = [
+  {
+    imdbID: "tt0468569",
+    title: "The Dark Knight",
+    year: "2008",
+    poster:
+      "https://m.media-amazon.com/images/M/MV5BMTMxNTMwODM0NF5BMl5BanBnXkFtZTcwODAyMTk2Mw@@._V1_.jpg",
+    runtime: 152,
+    imdbRating: 9.0,
+    userRating: 9.9,
+  },
+  {
+    imdbID: "tt1745960",
+    title: "Top Gun: Maverick",
+    year: "2022",
+    poster:
+      "https://m.media-amazon.com/images/M/MV5BZWYzOGEwNTgtNWU3NS00ZTQ0LWJkODUtMmVhMjIwMjA1ZmQwXkEyXkFqcGdeQXVyMjkwOTAyMDU@._V1_.jpg",
+    runtime: 130,
+    imdbRating: 8.2,
+    userRating: 9,
+  },
+];
+
+export default tempWatchedData;
